Integrate Dashboard

+++ client/www/components/dash/explorer/Explorer.tsx
@@ -1,7 +1,9 @@
 import { id, tx } from '@instantdb/core';
 import { InstantReactWebDatabase } from '@instantdb/react';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { isObject, debounce, last } from 'lodash';
+import { useCallback, useEffect, useMemo, useRef, useState, useContext } from 'react';
+import { jsonFetch } from '@/lib/fetch';
+import config from '@/lib/config';
 import produce from 'immer';
 import clsx from 'clsx';
 import CopyToClipboard from 'react-copy-to-clipboard';
@@ -23,7 +25,7 @@ import {
 } from '@heroicons/react/solid';
 import { PencilAltIcon } from '@heroicons/react/outline';
 
-import { errorToast } from '@/lib/toast';
+import { successToast, errorToast } from '@/lib/toast';
 import {
   ActionButton,
   ActionForm,
@@ -47,6 +49,7 @@ import {
   useNamespacesQuery,
   SearchFilter,
 } from '@/lib/hooks/explorer';
+import { TokenContext } from '@/lib/contexts';
 import { EditNamespaceDialog } from '@/components/dash/explorer/EditNamespaceDialog';
 import { EditRowDialog } from '@/components/dash/explorer/EditRowDialog';
 import { useRouter } from 'next/router';
@@ -148,7 +151,7 @@ function queryToFilters({
                 JSON.parse(part.value),
               ],
             ];
-          } catch (e) {}
+          } catch (e) { }
         }
         if (attr.checkedDataType === 'date') {
           try {
@@ -175,7 +178,7 @@ function queryToFilters({
                 opToInstaqlOp(part.operator),
                 JSON.parse(part.value),
               ]);
-            } catch (e) {}
+            } catch (e) { }
             break;
           }
           default: {
@@ -217,6 +220,11 @@ function sameFilters(
   return false;
 }
 
+const excludedSearchAttrs: [string, string][] = [
+  // Exclude computed fields
+  ["$files", "url"]
+];
+
 function SearchInput({
   onSearchChange,
   attrs,
@@ -257,9 +265,11 @@ function SearchInput({
   const comboOptions: { field: string; operator: string; display: string }[] = (
     attrs || []
   ).flatMap((a) => {
-    if (a.type === 'ref') {
+    const isExcluded = excludedSearchAttrs.some(([ns, name]) => ns === a.namespace && name === a.name);
+    if (a.type === 'ref' || isExcluded) {
       return [];
     }
+
     const ops = [];
 
     const opCandidates = [];
@@ -440,17 +450,17 @@ export function Explorer({
     [namespaces, currentNav?.namespace],
   );
 
-  const isSystemCatalogNs =
-    selectedNamespace != null &&
-    selectedNamespace.name != null &&
-    selectedNamespace.name.startsWith('$');
+  // auth
+  const token = useContext(TokenContext);
 
-  const readOnlyNs = isSystemCatalogNs && selectedNamespace.name !== '$users';
+  const isSystemCatalogNs = selectedNamespace?.name?.startsWith('$') ?? false;
+  const sanitizedNsName = selectedNamespace?.name ?? '';
+  const readOnlyNs = isSystemCatalogNs && !['$users', '$files'].includes(sanitizedNsName);
 
   const [limit, setLimit] = useState(50);
   const [offsets, setOffsets] = useState<{ [namespace: string]: number }>({});
 
-  const offset = offsets[selectedNamespace?.name ?? ''] || 0;
+  const offset = offsets[sanitizedNsName] || 0;
 
   const sortAttr = currentNav?.sortAttr || 'serverCreatedAt';
   const sortAsc = currentNav?.sortAsc ?? true;
@@ -495,7 +505,42 @@ export function Explorer({
     () => allItems.find((i) => i.id === editableRowId),
     [allItems.length, editableRowId],
   );
-  const rowText = Object.keys(checkedIds).length === 1 ? 'row' : 'rows';
+  const rowText =
+    sanitizedNsName === '$files'
+      ? Object.keys(checkedIds).length === 1 ? 'file' : 'files'
+      : Object.keys(checkedIds).length === 1 ? 'row' : 'rows';
+
+  // Storage
+
+  const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
+  const [uploadingFile, setUploadingFile] = useState(false);
+  const [customPath, setCustomPath] = useState('');
+  const fileInputRef = useRef<HTMLInputElement>(null);
+  const handleUploadFile = async () => {
+    try {
+      setUploadingFile(true);
+      if (selectedFiles.length === 0) {
+        return;
+      }
+
+      const [file] = selectedFiles;
+      const success = await upload(token, appId, file, customPath);
+
+      if (success) {
+        setSelectedFiles([]);
+        setCustomPath('');
+        fileInputRef.current && (fileInputRef.current.value = '');
+      }
+
+      // await refreshFiles();
+      successToast('Successfully uploaded!');
+    } catch (err: any) {
+      console.error('Failed to upload:', err);
+      errorToast(`('Failed to upload: ${err.body.message}`);
+    } finally {
+      setUploadingFile(false);
+    }
+  };
 
   return (
     <div className="relative flex w-full flex-1 overflow-hidden">
@@ -525,11 +570,16 @@ export function Explorer({
               }
               onClick={async () => {
                 try {
-                  await db.transact(
-                    Object.keys(checkedIds).map((id) =>
-                      tx[selectedNamespace.name][id].delete(),
-                    ),
-                  );
+                  if (selectedNamespace.name === "$files") {
+                    const filenames = allItems.filter(i => i.id in checkedIds).map(i => i.path as string);
+                    await bulkDeleteFiles(token, appId, filenames)
+                  } else {
+                    await db.transact(
+                      Object.keys(checkedIds).map((id) =>
+                        tx[selectedNamespace.name][id].delete(),
+                      ),
+                    );
+                  }
                 } catch (error) {
                   errorToast(`Failed to delete ${rowText}`);
                   return;
@@ -732,22 +782,65 @@ export function Explorer({
               </div>
             </div>
           </div>
+          {selectedNamespace.name === "$files"
+            ?
+            <div className="flex py-2 px-2 gap-2">
+              <div className="flex gap-2 w-full">
+                <div className="flex gap-2 flex-shrink-0">
+                  <input
+                    ref={fileInputRef}
+                    type="file"
+                    className="flex rounded-sm border border-zinc-200 bg-transparent px-1 py-1 text-sm shadow-sm transition-colors file:text-sm placeholder:text-zinc-500 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-zinc-950 disabled:cursor-not-allowed disabled:opacity-50"
+                    onChange={(e: React.ChangeEvent<any>) => {
+                      const files = e.target.files;
+                      setSelectedFiles(files);
+                      if (files?.[0]) {
+                        setCustomPath(files[0].name);
+                      }
+                    }}
+                  />
+                  <Button
+                    variant="primary"
+                    disabled={selectedFiles.length === 0}
+                    size="mini"
+                    loading={uploadingFile}
+                    onClick={handleUploadFile}
+                  >
+                    {uploadingFile ? "Uploading..." : "Upload file"}
+                  </Button>
+                </div>
+                <div className="relative flex flex-1 max-w-[67vw] min-w-0">
+                  <span className="absolute inset-y-0 left-0 flex items-center px-3 text-sm text-zinc-500 bg-gray-100 rounded-l-md ">File Path:</span>
+                  <input
+                    type="text"
+                    placeholder="Enter a custom path (optional)"
+                    value={customPath}
+                    onChange={(e) => setCustomPath(e.target.value)}
+                    className="w-full h-9 rounded-md bg-transparent pl-24 pr-3 py-1 text-sm placeholder:text-zinc-500 outline outline-1 outline-zinc-200 focus:ring-2 focus:ring-blue-700 border-0"
+                  />
+                </div>
+              </div>
+            </div>
+            : null}
           <div className="flex items-center justify-start space-x-2 p-1 text-xs border-b">
-            <Button
-              disabled={readOnlyNs}
-              title={
-                readOnlyNs
-                  ? `The ${selectedNamespace?.name} namespace is read-only.`
-                  : undefined
-              }
-              size="mini"
-              variant="secondary"
-              onClick={() => {
-                setAddItemDialogOpen(true);
-              }}
-            >
-              Add row
-            </Button>
+            {selectedNamespace.name !== "$files"
+              ?
+              <Button
+                disabled={readOnlyNs}
+                title={
+                  readOnlyNs
+                    ? `The ${selectedNamespace?.name} namespace is read-only.`
+                    : undefined
+                }
+                size="mini"
+                variant="secondary"
+                onClick={() => {
+                  setAddItemDialogOpen(true);
+                }}
+              >
+                Add row
+              </Button>
+              : null}
             <div>
               <Select
                 className="text-xs"
@@ -908,32 +1001,32 @@ export function Explorer({
                       onClick={
                         attr.sortable
                           ? () => {
-                              replaceNavStackTop({
-                                sortAttr: attr.name,
-                                sortAsc:
-                                  sortAttr !== attr.name ? true : !sortAsc,
-                              });
-                            }
+                            replaceNavStackTop({
+                              sortAttr: attr.name,
+                              sortAsc:
+                                sortAttr !== attr.name ? true : !sortAsc,
+                            });
+                          }
                           : attr.name === 'id'
                             ? () => {
-                                replaceNavStackTop({
-                                  sortAttr: 'serverCreatedAt',
-                                  sortAsc:
-                                    sortAttr !== 'serverCreatedAt'
-                                      ? true
-                                      : !sortAsc,
-                                });
-                              }
+                              replaceNavStackTop({
+                                sortAttr: 'serverCreatedAt',
+                                sortAsc:
+                                  sortAttr !== 'serverCreatedAt'
+                                    ? true
+                                    : !sortAsc,
+                              });
+                            }
                             : undefined
                       }
                     >
                       <div className="flex items-center gap-2">
-                        {attr.name}
+                        {(selectedNamespace.name === "$files" && attr.name === "url") ? "" : attr.name}
                         {attr.sortable || attr.name === 'id' ? (
                           <span>
                             {sortAttr === attr.name ||
-                            (sortAttr === 'serverCreatedAt' &&
-                              attr.name === 'id') ? (
+                              (sortAttr === 'serverCreatedAt' &&
+                                attr.name === 'id') ? (
                               sortAsc ? (
                                 '↓'
                               ) : (
@@ -973,7 +1066,7 @@ export function Explorer({
                           );
                         }}
                       />
-                      {readOnlyNs ? null : (
+                      {(readOnlyNs || sanitizedNsName === "$files") ? null : (
                         <button
                           className="opacity-0 group-hover:opacity-100 transition-opacity"
                           onClick={() => setEditableRowId(item.id)}
@@ -993,22 +1086,31 @@ export function Explorer({
                               : '80px',
                         }}
                       >
-                        <ExplorerItemVal
-                          item={item}
-                          attr={attr}
-                          onClickLink={() => {
-                            const linkConfigDir =
-                              attr.linkConfig[
+                        {selectedNamespace.name === "$files" && attr.name === "url"
+                          ? <Button
+                            variant='secondary'
+                            size='mini'
+                            onClick={() => {
+                              window.open(item.url as string, '_blank');
+                            }}>
+                            View File
+                          </Button>
+                          : <ExplorerItemVal
+                            item={item}
+                            attr={attr}
+                            onClickLink={() => {
+                              const linkConfigDir =
+                                attr.linkConfig[
                                 !attr.isForward ? 'forward' : 'reverse'
-                              ];
-                            if (linkConfigDir) {
-                              pushNavStack({
-                                namespace: linkConfigDir.namespace,
-                                where: [`${linkConfigDir.attr}.id`, item.id],
-                              });
-                            }
-                          }}
-                        />
+                                ];
+                              if (linkConfigDir) {
+                                pushNavStack({
+                                  namespace: linkConfigDir.namespace,
+                                  where: [`${linkConfigDir.attr}.id`, item.id],
+                                });
+                              }
+                            }}
+                          />}
                       </td>
                     ))}
                   </tr>
@@ -1225,4 +1327,50 @@ function _dev(db: InstantReactWebDatabase<any>) {
     };
     (window as any).i = i;
   }
+}
+
+// Storage
+
+async function upload(
+  token: string,
+  appId: string,
+  file: File,
+  customFilename: string
+): Promise<boolean> {
+  const formData = new FormData();
+  formData.append('file', file);
+  formData.append('path', customFilename || file.name);
+
+  const data = await jsonFetch(
+    `${config.apiURI}/dash/apps/${appId}/storage/upload`,
+    {
+      method: 'POST',
+      headers: {
+        authorization: `Bearer ${token}`,
+      },
+      body: formData,
+    }
+  );
+
+  return data;
+}
+
+async function bulkDeleteFiles(
+  token: string,
+  appId: string,
+  filenames: string[]
+): Promise<any> {
+  const { data } = await jsonFetch(
+    `${config.apiURI}/dash/apps/${appId}/storage/files/delete`,
+    {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+        authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify({ filenames }),
+    }
+  );
+
+  return data;
 }
